Move Biome REST API tests to appropriate REST APIs
Moves the existing Biome REST API tests to their new locations
(credentials and key management). The key management tests require
credentials to be enabled because they use the request's authentication
(e.g. Biome user).
